refactor(exceptions): move HTTP exceptions to src/Exception/Http/
Move HTTP exception classes from src/Common/Http/Exception/ to src/Exception/Http/
for consistency with the core exception types in src/Exception/.
