+++ src/Exception/Http/HttpExceptionInterface.php
@@ -13,7 +13,7 @@
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-namespace OpenEMR\Common\Http\Exception;
+namespace OpenEMR\Exception\Http;
 
 interface HttpExceptionInterface extends \Throwable
 {
