+++ src/Exception/Http/NotFoundHttpException.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Exception for HTTP 500 Internal Server Error
+ * Exception for HTTP 404 Not Found
  *
  * @package   OpenEMR
  * @link      http://www.open-emr.org
@@ -10,11 +10,11 @@
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-namespace OpenEMR\Common\Http\Exception;
+namespace OpenEMR\Exception\Http;
 
 use OpenEMR\Common\Http\StatusCode;
 
-class InternalServerErrorHttpException extends HttpException
+class NotFoundHttpException extends HttpException
 {
     /**
      * @param array<string, string|string[]> $headers
@@ -24,6 +24,6 @@ class InternalServerErrorHttpException extends HttpException
         ?\Throwable $previous = null,
         array $headers = []
     ) {
-        parent::__construct(StatusCode::INTERNAL_SERVER_ERROR, $message, $previous, $headers);
+        parent::__construct(StatusCode::NOT_FOUND, $message, $previous, $headers);
     }
 }
