+++ src/Exception/Http/UnauthorizedHttpException.php
@@ -10,7 +10,7 @@
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-namespace OpenEMR\Common\Http\Exception;
+namespace OpenEMR\Exception\Http;
 
 use OpenEMR\Common\Http\StatusCode;
